third_party/blink/web_tests/wpt_internal/app-history/navigate-event/navigate-window-open-self.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(t => {
  appHistory.onnavigate = t.step_func_done(e => {
    assert_equals(e.canRespond, true);
    assert_equals(e.hashChange, true);
    assert_equals(e.formData, null);
    e.preventDefault();
  });
  window.onload = t.step_func(() => window.open("#", "_self"));
}, "window.open() fires navigate event when targeting self");
</script>
